+++ apps/studio/components/interfaces/QueryPerformance/QueryPerformanceLoadingRow.tsx
@@ -1,0 +1,17 @@
+import ShimmeringLoader from 'ui-patterns/ShimmeringLoader'
+
+export const QueryPerformanceLoadingRow = ({ colSpan }: { colSpan: number }) => {
+  return (
+    <>
+      {Array(4)
+        .fill('')
+        .map((_, i) => (
+          <tr key={'loading-' + { i }}>
+            <td colSpan={colSpan}>
+              <ShimmeringLoader />
+            </td>
+          </tr>
+        ))}
+    </>
+  )
+}
